+++ lib/representable/xml/collection.rb
@@ -11,7 +11,7 @@ module Representable::XML
 
     module Methods
       def create_representation_with(doc, options, format)
-        bin = representable_mapper(format, options).bindings(represented, options).first
+        bin = representable_mapper(format, options).bindings.first
         bin.write(doc, super, bin.name)
       end